requirements.txt
Flask
Flask-JWT-Extended
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
requests
psycopg2-binary
gunicorn
